Fixed Hibernate issue with authorities saving part.2
