+++ src/main/java/com/zufar/onlinestore/user/api/UserService.java
@@ -6,6 +6,7 @@ import com.zufar.onlinestore.user.entity.UserEntity;
 import com.zufar.onlinestore.user.exception.UserAlreadyRegisteredException;
 import com.zufar.onlinestore.user.exception.UserNotFoundException;
 import com.zufar.onlinestore.user.repository.UserRepository;
+import com.zufar.onlinestore.user.validation.UserDataValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -22,27 +23,12 @@ public class UserService implements UserApi {
     private final UserRepository userCrudRepository;
     private final UserDtoConverter userDtoConverter;
     private final AuthorityService authorityService;
+    private final UserDataValidator userDataValidator;
 
     @Override
     public UserDto saveUser(final UserDto userDto) {
         UserEntity userEntity = userDtoConverter.toEntity(userDto);
-
-        List<String> errors = new ArrayList<>();
-
-        if (!isEmailUnique(userEntity.getUserId(), userEntity.getEmail())) {
-            errors.add(String.format("User with email = %s is already registered", userEntity.getEmail()));
-        }
-
-        if (!isUsernameUnique(userEntity.getUserId(), userEntity.getUsername())) {
-            errors.add(String.format("User with username = %s is already registered", userEntity.getUsername()));
-        }
-
-        if (!errors.isEmpty()) {
-
-            throw new UserAlreadyRegisteredException(errors);
-
-        }
-
+        userDataValidator.validate(userEntity);
         UserEntity userEntityWithId = userCrudRepository.save(userEntity);
         authorityService.setDefaultAuthority(userEntityWithId);
 
@@ -59,27 +45,4 @@ public class UserService implements UserApi {
         return userDtoConverter.toDto(userEntity.get());
     }
 
-    public boolean isEmailUnique(UUID id, String email) {
-        Optional<UserEntity> userByEmail = userCrudRepository.findByEmail(email);
-
-        if (userByEmail.isEmpty()) {
-            return true;
-        }
-
-        return id == null || userByEmail.map(UserEntity::getUserId).filter(userId -> userId == id).isPresent();
-    }
-
-    public boolean isUsernameUnique(UUID id, String username) {
-        UserEntity userByUsername = userCrudRepository.findUserByUsername(username);
-
-        if (userByUsername == null) {
-            return true;
-        }
-
-        if (id == null) {
-            return false;
-        } else {
-            return userByUsername.getUserId() == id;
-        }
-    }
 }
